hisi_sas: fix p660_hard_phy_reset race condition
We have a condition where the phyup occurs before
the port is deformed, so so try to re-form a form port
(which fails), and then defrom the port.

It seems to take 200ms between
notify_phy_event(sas_phy, PHYE_LOSS_OF_SIGNAL) being called
and the port actually being deformed. This is bad as the phy
is down and up before this 200ms; I cannot sleep between
these events as p660_hard_phy_reset is called from the same
context as port deformed. So I set a timer.

This needs to be revisited.

Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: flyingnosky <flyingnosky@163.com>