{fromtree} scsi: hisi_sas: stub hisi_sas_port_deformed()
Currently when a root PHY is deformed from a root
port then we try to release the slots.

But it is not right to release this early. It is actually
safer to release slots when instructed to do so my the
upper layer.

This patch stubs the deformed function. It would be
nice to actually remove hisi_sas_port_{de}formed()
function pair, however we cannot as we need to know the
asd_sas_port libsas has associated with the asd_sas_phy.
The hw does actually generate a port id, but this seems
to a random number generator.

Signed-off-by: John Garry <john.garry@huawei.com>
Reviewed-by: Xiang Chen <chenxiang66@hisilicon.com>