{fromtree} scsi: hisi_sas: modify hisi_sas_abort_task() for SATA
In the case of successfully aborting a SATA command, we
mark all the commands in the target device as aborted.

Also remove call to nexus and controller reset, as if the
softreset fails, the upper-layer will recognise this
and escalate.

Note: we don't check the return from hisi_sas_internal_task_abort().
Even though it may tell us the command was aborted and we no
longer need to issue the softreset, however it is safer to still
try the softreset.

Signed-off-by: John Garry <john.garry@huawei.com>
Reviewed-by: Xiang Chen <chenxiang66@hisilicon.com>