{topost} scsi: hisi_sas: make upper layer deal with SSP erroneous remnant IO
For SSP slots which complete with error and maybe have IO still in
target device, allow them to timeout so upper layer can deal
with them.

Previously, we should try to abort the IO in the LLDD and then
pass back to upper layer, but sometimes this would also error.
Instead of adding special error handling for this scenario in
the LLDD, allow the upper layer to handle completely.

No performance hit is seen by taking this approach.

Signed-off-by: John Garry <john.garry@huawei.com>