{topost} hisi_sas: add v2 hw internal abort timeout workaround
This patch work around a soc bug. In v2 hw, channel should become idle
in order to finish abort process. If target side has been sending HOLD,
host side channel failed to complete the frame to send, and can not
enter the idle state. Then internal abort command will timeout.

As this issue is only in v2 hw, We deal with it in the hw layer.
Our workaround solution is:
If abort is not finished within a certain period of time, we will check
HOLD status. If HOLD has been sending, we will send break command.

Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>