{fromtree} hisi_sas: release SMP slot in lldd_abort_task when SMP IO is not found in SAS controller
When SMP IO timeout, it will call lldd_abort_task to release slot of SMP IO, and
then will release sas_task of SMP IO. In lldd_abort_task, we send internal abort
to sigle SMP IO, if SMP IO is not in SAS controller, slot of SMP IO is not released,
but sas_task will still be released, so it exists slot's sas_task is NULL
(slot->task=NULL) which will cause NULL pointer when hisi_sas_slot_task_free happens.

So check the return value of internal abort for SMP IO, if SMP IO is not in SAS
controller,release slot of SMP IO in lldd_abort_task.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Reviewed-by: John Garry <john.garry@huawei.com>