{topost} net: hns3: reinitializing bd information when getting a new rx buffer
If valid bit of bd information is set, this bit is always valid whether
HW has set it or has not set it. Because we do not clear it, when getting
a new rx buffer. It will cause a problem that some packets whose valid bit
are not set by HW are received by SW, because valid bit of bd information
is used by SW to check if the bd is valid. Although bd information will be
rewrited by HW, bd information needs to be reinitialized when getting a
new rx buffer.

Signed-off-by: Fuyun Liang <liangfuyun1@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>