{topost} net: hns3: clear TX/RX rings when stopping port
When we down the port, some packets are left in TX/RX buffer. When we
up the port again, these old packets are forwarded to protocol stack
or are sent to internet. It will make some problem. TX/RX buffer should
be cleared when stopping port. This patch adds some function to ensure
the buffer is clean when port is started.

Meanwhile, some packets may are being handled by SW when we have called
the stop function. It is not suitable. Vectors should be disabled before
calling stop function.

Signed-off-by: Fuyun Liang <liangfuyun1@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>